requirements.txt
camelot-py[cv]
pdfplumber
pandas 